refactored CK UI. Check for user before adding works
